Add selector comments

+++ test/safelog.t.sol
@@ -7,17 +7,17 @@ import {safelog} from "src/safelog.sol";
 /// @author philogy <https://github.com/philogy>
 contract safelogTest is Test {
     function testLog() public {
-        uint[] memory wordsBefore = new uint[](32);
-        for (uint i; i < wordsBefore.length; i++) {
-            uint word;
+        uint256[] memory wordsBefore = new uint[](32);
+        for (uint256 i; i < wordsBefore.length; i++) {
+            uint256 word;
             assembly {
                 word := mload(mul(i, 0x20))
             }
             wordsBefore[i] = word;
         }
         safelog.log("hello (%s)", 30);
-        for (uint i; i < wordsBefore.length; i++) {
-            uint word;
+        for (uint256 i; i < wordsBefore.length; i++) {
+            uint256 word;
             assembly {
                 word := mload(mul(i, 0x20))
             }
